StringDifferenceAnalyzer does not stop a first minor difference
