+++ NFluent.35/NFluent.35.csproj
@@ -172,6 +172,9 @@
     <Compile Include="..\NFluent.Core\Helpers\IsInstanceHelper.cs">
       <Link>Helpers\IsInstanceHelper.cs</Link>
     </Compile>
+    <Compile Include="..\NFluent.Core\Helpers\StringDifferenceAnalyzer.cs">
+      <Link>Helpers\StringDifferenceAnalyzer.cs</Link>
+    </Compile>
     <Compile Include="..\NFluent.Core\Helpers\TimeHelper.cs">
       <Link>Helpers\TimeHelper.cs</Link>
     </Compile>
