+++ NFluent.PCL/NFluent.PCL.csproj
@@ -172,6 +172,9 @@
     <Compile Include="..\NFluent.Core\Helpers\IsInstanceHelper.cs">
       <Link>Helpers\IsInstanceHelper.cs</Link>
     </Compile>
+    <Compile Include="..\NFluent.Core\Helpers\StringDifferenceAnalyzer.cs">
+      <Link>Helpers\StringDifferenceAnalyzer.cs</Link>
+    </Compile>
     <Compile Include="..\NFluent.Core\Helpers\TimeHelper.cs">
       <Link>Helpers\TimeHelper.cs</Link>
     </Compile>
@@ -343,7 +346,7 @@
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <ProjectExtensions>
     <VisualStudio>
-      <UserProperties ProjectLinkReference="3a474510-f384-41ce-9270-539b1e51da7e" ProjectLinkerExcludeFilter="\\?desktop(\\.*)?$;\\?silverlight(\\.*)?$;\.desktop;\.silverlight;\.xaml;^service references(\\.*)?$;\.clientconfig;^web references(\\.*)?$" />
+      <UserProperties ProjectLinkerExcludeFilter="\\?desktop(\\.*)?$;\\?silverlight(\\.*)?$;\.desktop;\.silverlight;\.xaml;^service references(\\.*)?$;\.clientconfig;^web references(\\.*)?$" ProjectLinkReference="3a474510-f384-41ce-9270-539b1e51da7e" />
     </VisualStudio>
   </ProjectExtensions>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
