+++ tests/Nfluent.Core.Internals.Tests/ForDocumentation/CheckList.cs
@@ -19,7 +19,6 @@
 namespace NFluent.Tests.ForDocumentation
 {
     using System.Collections.Generic;
-    using System.Xml;
     using System.Xml.Serialization;
     
     /// <summary>
